Split execute.ts into src/execute/, one function per file
The 572-code-line module becomes a folder: each function in its own file,
runCheckout and runSetupNode lifted out of runUses, and an index.ts bucket
carrying the module doc and the same public surface. The 1300-line test file
carves along the same seams, with compact colocated suites for the pieces the
integration suite reaches only indirectly.
