+++ src/index.ts
@@ -4,7 +4,7 @@ export { expandMatrix } from "./matrix/index.js";
 export { evalIf, expandWorkflowJobs } from "./jobs/index.js";
 export { parseUses } from "./uses/index.js";
 export { makeOctokit, predict } from "./predict/index.js";
-export type { ExecOutcome, JobExecutor } from "./execute.js";
+export type { ExecOutcome, JobExecutor } from "./execute/index.js";
 export type {
   JobName,
   WorkflowEntry,
